Use lookbehind to ensure quote is not escaped
The problem was code like this:

```json
{
  "foo": "{\"bar\": \"qux\"}"
  "_":  "└────────┘ <- jsonKeywordMatch (!)"
}
```

before, we'd highlight the value of the "foo" key as if it were JSON,
because `jsonKeywordMatch` would match against the region shown above.

By adding the `\\\@<!` before the `"` we ensure that a `\` doesn't come
before the `"`; a keyword must be terminated by an unescaped quote.